Latest version changed by Thanh

+++ src/agent/UniformVsDefender.java
@@ -38,8 +38,8 @@ public final class UniformVsDefender extends Defender {
 	private int numAttActionSample = DEFAULT_NUM_STATE_SAMPLE; // number of attack actions to sample
 	
 	public UniformVsDefender(final double logisParam, final double discFact, final double thres
-		, final int maxNumRes, final int minNumRes, final double numResRatio
-		, final int maxNumSelectACandidate, final int minNumSelectACandidate, final double numSelectACandidateRatio) {
+		, final double maxNumRes, final double minNumRes, final double numResRatio
+		, final double maxNumSelectACandidate, final double minNumSelectACandidate, final double numSelectACandidateRatio) {
 		super(DefenderType.vsUNIFORM);
 		if (discFact <= 0.0 || discFact > 1.0 || thres < 0.0 || thres > 1.0
 			|| minNumRes < 1 || maxNumRes < minNumRes || numResRatio < 0.0 || numResRatio > 1.0
@@ -52,12 +52,12 @@ public final class UniformVsDefender extends Defender {
 		this.discFact = discFact;
 		this.thres = thres;
 		
-		this.maxNumRes = maxNumRes;
-		this.minNumRes = minNumRes;
-		this.numResRatio = numResRatio;
+		this.maxNumRes = (int)maxNumRes;
+		this.minNumRes = (int)minNumRes;
+		this.numResRatio = (int)numResRatio;
 		
-		this.maxNumSelectACandidate = maxNumSelectACandidate;
-		this.minNumSelectACandidate = minNumSelectACandidate;
+		this.maxNumSelectACandidate = (int)maxNumSelectACandidate;
+		this.minNumSelectACandidate = (int)minNumSelectACandidate;
 		this.numSelectACandidateRatio = numSelectACandidateRatio;
 	}
